+++ public/index.html
@@ -25,6 +25,37 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <title>SESI SENAI - Validação de atividades complementares</title>
+    <style>
+      .hero.welcome.is-info {
+        background: #36D1DC;
+        background: -webkit-linear-gradient(to right, #5B86E5, #36D1DC);
+        background: linear-gradient(to right, #0699a2, #36D1DC);
+      }
+      .navbar .container {
+        margin-left: 30px;
+      }
+      @media screen and (max-width: 1060px) {
+        nav, .navbar-menu {
+            display: block !important;
+        }
+      }
+      .section {
+        padding: 1rem 1.5rem !important;
+      }
+      .is-centered  {
+        text-align: center;
+      }
+      @media screen and (max-width: 1024px){
+          .hidden-mobile {
+            display: none !important;
+          }
+      }
+      @media screen and (min-width: 1024px){
+        .hidden-web {
+          display: none !important;
+        }
+      }
+    </style>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
